+++ src/libse/Http/DownloaderFactory.cs
@@ -2,6 +2,7 @@
 using System.Net;
 using System.Net.Http;
 using Nikse.SubtitleEdit.Core.Common;
+using Nikse.SubtitleEdit.Core.Settings;
 
 namespace Nikse.SubtitleEdit.Core.Http
 {
