+++ src/libse/Settings/BeautifyTimeCodesSettings.cs
@@ -1,4 +1,4 @@
-namespace Nikse.SubtitleEdit.Core.Common
+namespace Nikse.SubtitleEdit.Core.Settings
 {
     public class BeautifyTimeCodesSettings
     {
