+++ src/libse/Settings/FixCommonErrorsSettings.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using System.Text;
 
-namespace Nikse.SubtitleEdit.Core.Common
+namespace Nikse.SubtitleEdit.Core.Settings
 {
     public class FixCommonErrorsSettings
     {
