+++ src/libse/Settings/MultipleSearchAndReplaceGroup.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Nikse.SubtitleEdit.Core.Common
+namespace Nikse.SubtitleEdit.Core.Settings
 {
     public class MultipleSearchAndReplaceGroup
     {
