+++ src/libse/Settings/RecentFileEntry.cs
@@ -1,4 +1,4 @@
-namespace Nikse.SubtitleEdit.Core.Common
+namespace Nikse.SubtitleEdit.Core.Settings
 {
     public class RecentFileEntry
     {
