+++ src/libse/Settings/SubtitleBeaming.cs
@@ -1,6 +1,6 @@
 using System.Drawing;
 
-namespace Nikse.SubtitleEdit.Core.Common
+namespace Nikse.SubtitleEdit.Core.Settings
 {
     public class SubtitleBeaming
     {
