+++ src/ui/Forms/BeautifyTimeCodes/BeautifyTimeCodesProfile.cs
@@ -2,6 +2,7 @@
 using System.Drawing;
 using System.Windows.Forms;
 using Nikse.SubtitleEdit.Core.Common;
+using Nikse.SubtitleEdit.Core.Settings;
 using Nikse.SubtitleEdit.Logic;
 
 namespace Nikse.SubtitleEdit.Forms.BeautifyTimeCodes
