+++ src/ui/Forms/BeautifyTimeCodes/BeautifyTimeCodesProfileSimple.cs
@@ -4,6 +4,7 @@ using Nikse.SubtitleEdit.Core.SubtitleFormats;
 using Nikse.SubtitleEdit.Logic;
 using System;
 using System.Windows.Forms;
+using Nikse.SubtitleEdit.Core.Settings;
 
 namespace Nikse.SubtitleEdit.Forms.BeautifyTimeCodes
 {
